# Avoid unnecessary string allocations during memory-idx.Add