This is a small optimization (shaved about 3% off the heap size in our POC cluster, not major). 

Instead of splitting and joining, the code iterates over the string in place and just saves references to the pieces of the string.

Benchmark:
```
benchmark               old ns/op     new ns/op     delta
BenchmarkIndexing-8     7607          5287          -30.50%

benchmark               old allocs     new allocs     delta
BenchmarkIndexing-8     22             18             -18.18%

benchmark               old bytes     new bytes     delta
BenchmarkIndexing-8     1246          1150          -7.70%
```