@Dieterbe 2017-07-17
I suspect the `log.Debug` calls are somehow interfering. So I did a new bench run with all log.Debug calls commented out in both master branch and your branch. I also found a cleaner way to run only the benchmark we care about, like so:
```
$ go test -run='^$' -bench=BenchmarkIndexing -count 20 >/dev/null; go test -run='^$' -bench=BenchmarkIndexing -count 20 > new2.txt; git checkout master; go test -run='^$' -bench=BenchmarkIndexing -count 20 > master2.txt


$ benchstat master2.txt new2.txt
name        old time/op    new time/op    delta
Indexing-8    3.74µs ±10%    3.61µs ± 8%   -3.61%  (p=0.040 n=19+20)

name        old alloc/op   new alloc/op   delta
Indexing-8    1.19kB ± 0%    1.13kB ± 0%   -5.20%  (p=0.000 n=19+17)

name        old allocs/op  new allocs/op  delta
Indexing-8      16.0 ± 0%      14.0 ± 0%  -12.50%  (p=0.000 n=20+20)
```

so this looks a lot better.
optimizng the log.Debug calls (eg to not execute any logic when Debug is not enabled) is on the roadmap for the future, so we don't have to worry about it here.